+++ src/tools/metrics/mod.rs
@@ -2,4 +2,3 @@
 
 pub mod recorder;
 pub mod tables;
-pub mod summary;
